Not allow to zoom in the internal pages.
This commit prevents an user from zooming in the Home Page, Bookmarks
Page and History Page.
On some devices, for examlpe, some tablets with low dpi, tapping the
search bar in the Home Page would zoom the browser in about 25% and then
open the keyboard. That looks weird.
Also, the Lightning Browser does not allow zooming in the Settings menu
and the Tabs/Bookmarks sidebars. Allowing the user to zoom in one place
of the UI and not in the other is inconsitent.