respond to generic panic trigger Intent by quitting
PanicKit provides a common framework for creating "panic button" apps that
can trigger actions in "panic responder" apps.  In this case, the response
is to close the app.  That way, if the user has configured various things
to be cleared when the app closes, pressing the panic button will make all
those things be cleared.

For more info:
https://github.com/guardianproject/panickit
https://dev.guardianproject.info/projects/panic/wiki

Conflicts:
	app/src/main/java/acr/browser/lightning/activity/BrowserActivity.java